+++ changelog.md
@@ -1,5 +1,7 @@
 ## Forestry 2.2.4
 - Fix saplings with default genomes not generating leaves properly (#144)
+- Add API for IBeekeepingLogic to change throttle ticks
+- Fix NBT loading bug in BeekeepingLogic to load queen max health
 
 ## Forestry 2.2.3
 - Add Cathemeral bee activity type, which is a random activity period of 12000 ticks that varies based on position. Check times using /forestry bee cathemeral